Make read handle receiving less-than-requested from obj-server
In the course of adding this "fix" to the problem of the server returning
less than expected, I realized that what was really going on was that the
computation of how much to expect was flawed.  After this fix, we can
handle getting less than expected, but I don't observe that actually
happening anymore, in my quick test.  I think it's probably legit for a
server to return less than requested, so I'm leaving the fix in.

+++ fuse/src/common.c
@@ -1201,7 +1201,7 @@ ssize_t write_recoveryinfo(ObjectStream* os, const PathInfo* const info) {
    ///      memset(rec, 1, recovery); // stands out in objects written from /dev/zero
    ///      needs_init = 0;
    ///   }
-   static int dbg=0x11;
+   static uint8_t dbg=0x11;
    memset(rec, dbg, recovery); // stands out in objects written from /dev/zero
    dbg += 0x11;
